pyproject.toml
[project]
name = "timm"
version = "0.1.0"
description = "OKR micro-model via Agent-Specific Model Synthesis on Apple MLX"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlx>=0.22.0",
    "mlx-lm>=0.21.0",
    "sentencepiece>=0.2.0",
    "anthropic>=0.49.0",
    "pyyaml>=6.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
